# docs/swagger.yaml
definitions:
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        format: int64
        type: integer
    type: object
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
...
  wn_internal_domain_dto.ConfirmationCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
      newPassword:
        type: string
    required:
    - code
    - email
    type: object
...
  wn_internal_domain_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    type: object
  wn_internal_domain_dto.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      userId:
        type: string
    type: object
...
  wn_internal_domain_dto.RegisterCredentials:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  wn_internal_domain_dto.RegisterResponse:
    properties:
      userId:
        type: string
    type: object
  wn_internal_domain_dto.SendCodeResponse:
    properties:
      nextCodeDelay:
        $ref: '#/definitions/time.Duration'
    type: object
...
  wn_internal_domain_dto.UploadFileRequest:
    properties:
      file:
        $ref: '#/definitions/multipart.FileHeader'
    required:
    - file
    type: object
  wn_internal_domain_dto.UploadFileResponse:
    properties:
      imgUrl:
        type: string
    type: object
  wn_internal_domain_services_token.UserTokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    required:
    - accessToken
    - refreshToken
    type: object
  wn_pkg_response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/wn_pkg_response.meta'
      pagination:
        $ref: '#/definitions/wn_pkg_response.pagination'
    type: object
  wn_pkg_response.meta:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      requestId:
        type: string
    type: object
  wn_pkg_response.pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      perPage:
        type: integer
    type: object
info:
  contact: {}
  description: This is walrus notes api service.
  title: WALRUS NOTES API
  version: "1.0"
paths:
  /wn/api/v1/auth/code:
    post:
      description: register new user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.SendCodeResponse'
              type: object
        "400":
          description: 'possible codes: incorrect_password'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found, confirm_code_already_send'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: send_confirm_code
      tags:
      - auth
  /wn/api/v1/auth/confirm:
    post:
      description: Подтверждение кода для подтверждения почты, либо сброса пароля.
        Если сброс пароля, то newPassword обязательное поле.
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.ConfirmationCodeRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found, confirm_code_incorrect, confirm_code_not_exist,
            no_new_password'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: confirm_code
      tags:
      - auth
  /wn/api/v1/auth/forgot:
    post:
      description: Сброс пароля
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.SendCodeResponse'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found, confirm_code_already_send'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: forgot_password
      tags:
      - auth
  /wn/api/v1/auth/login:
    post:
      description: Получение access,refresh токенов по почте и паролю
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.LoginResponse'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "401":
          description: 'possible codes: incorrect_password'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found '
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: login
      tags:
      - auth
  /wn/api/v1/auth/refresh:
    post:
      description: Получение access,refresh токенов по access, refresh токенам
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_services_token.UserTokens'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_services_token.UserTokens'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: bad_refresh_token, bad_access_token, bad_token_claims,
            token_dont_exist, tokens_dont_match'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: refresh_tokens
      tags:
      - auth
  /wn/api/v1/auth/register:
    post:
      description: register new user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.RegisterCredentials'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.RegisterResponse'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: not_unique'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: register_user
      tags:
      - auth
  /wn/api/v1/file/upload:
    post:
      description: загрузить файл
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.UploadFileRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.UploadFileResponse'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: upload_file
      tags:
      - file
...
swagger: "2.0"